[5/7] cleanups: various small changes
Like the previous patch, this patch groups several small changes. But the
changes in this patch do slightly change (improve) the meaning of the code.

+++ src/CartridgeSlotManager.cc
@@ -365,7 +365,7 @@ string CartCmd::execute(const vector<string>& tokens, EmuTime::param /*time*/)
 		string slotname = (cartname.size() == 5)
 			? string(1, cartname[4])
 			: "any";
-		int extensionNameToken = 1;
+		size_t extensionNameToken = 1;
 		if (tokens[1] == "insert") {
 			if (tokens.size() > 2) {
 				extensionNameToken = 2;
